refector: teamArray 컴포 분리
